Convert to Typescript
