+++ package.json
@@ -2,10 +2,10 @@
   "name": "essential",
   "version": "1.0.0",
   "description": "Star Inc. Web研發技能訓練教學 ",
-  "main": "app.js",
+  "main": "app.ts",
   "scripts": {
-    "serve": "node ./app.js",
-    "dev-serve": "node_modules/.bin/nodemon ./app.js"
+    "dev": "node_modules/.bin/nodemon app.ts",
+    "serve": "node --loader ts-node/esm app.ts"
   },
   "keywords": [],
   "author": "Star Inc.",
@@ -13,10 +13,17 @@
   "dependencies": {
     "express": "^4.17.1",
     "http-status-codes": "^2.1.4",
-    "sequelize": "^6.7.0",
-    "sqlite3": "^5.0.2"
+    "reflect-metadata": "^0.1.13",
+    "sequelize": "^6.12.2",
+    "sequelize-typescript": "^2.1.1",
+    "sqlite3": "^5.0.2",
+    "ts-node": "^10.4.0"
   },
   "devDependencies": {
-    "nodemon": "^2.0.13"
+    "@types/express": "^4.17.13",
+    "@types/node": "^17.0.4",
+    "@types/validator": "^13.7.1",
+    "nodemon": "^2.0.13",
+    "typescript": "^4.5.4"
   }
 }
