+++ src/kernel/database.ts
@@ -1,0 +1,3 @@
+import {Sequelize} from 'sequelize-typescript';
+
+export default new Sequelize('sqlite::memory:');
