+++ src/models/article.ts
@@ -1,0 +1,10 @@
+import { Table, Column, Model } from 'sequelize-typescript';
+
+@Table
+export default class Article extends Model {
+    @Column
+    title: string
+
+    @Column
+    content: string
+}
